[REFACTOR] Simplify expense filtering by extracting time period logic into a separate function
