+++ src/app/services/user_service.py
@@ -7,9 +7,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from fastapi import status
 
 from app.schemas.user import UserRegistration, UserResponse
-from app.schemas.common.messages import ResponseMessage
-from app.services.utils import processors as p, validators as v, utils as u
+from app.schemas.common.common import ResponseMessage
 from app.schemas.common.application_error import ApplicationError
+from app.services.utils import processors as p, validators as v, utils as u
 from app.sql_app.user.user import User
 
 logger = logging.getLogger(__name__)
